% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{babyjustnames}
\alias{babyjustnames}
\title{Babyname data from the US and NZ}
\format{A data frame with 1883170 rows and 4 variables:
\describe{
  \item{Name}{The name of the baby}
  \item{Year}{The year the baby was born with that name}
  \item{Country}{The country: NZ or US}
  \item{Count}{The number of babies born with this name}
}}
\usage{
babyjustnames
}
\description{
Babyname data from the US and NZ
}
\keyword{datasets}
