% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{babynames}
\alias{babynames}
\title{Babyname data from the US and NZ}
\format{A data frame with 2057883 rows and 5 variables:
\describe{
  \item{Name}{The name of the baby}
  \item{Sex}{The recorded sex of the baby at birth}
  \item{Year}{The year the baby was born with that name}
  \item{Country}{The country: NZ or US}
  \item{Count}{The number of babies born with this name}
}}
\usage{
babynames
}
\description{
Babyname data from the US and NZ
}
\keyword{datasets}
